Move Tuxepedia SQLite interface stub to separate module
