+++ tuxemon/core/tuxepedia/extractor.py
@@ -5,27 +5,18 @@
     license: GPLv3
 """
 
-from contextlib import contextmanager
 import logging
-import sqlite3
 
 from lxml import html
 import requests
-
-from . import RESOURCE_PATHS
 
 
 class TuxepediaWebExtractor:
     """requests + lxml wrapper class to extract Tuxemon
     information from the Tuxepedia website"""
 
-    def __init__(self, web_path):
-        """
+    def __init__(self):
 
-        :param web_path: root path of the Tuxepedia website
-        """
-
-        self.web_path = web_path
         self.db_connection = None
 
         # TODO: add more Web params, like 'Content-Type' as needed
@@ -34,31 +25,6 @@ class TuxepediaWebExtractor:
     def get_logger(self):
         """Access a custom class logger"""
         return logging.getLogger(self.__class__.__name__)
-
-    @contextmanager
-    def db_connect(self):
-        """Connect to the SQLite database backend to store information"""
-
-        self.connection = sqlite3.connect(RESOURCE_PATHS.database)
-
-        yield self
-
-        self.connection.close()
-        self.connection = None
-
-    @contextmanager
-    def cursor(self):
-        """Yield a cursor for database backend queries"""
-
-        if self.connection is None:
-            error_msg = "Cursor for db {} could not be created"
-            error_msg = error_msg.format(RESOURCE_PATHS.database)
-
-            self.get_logger().error(error_msg)
-
-        yield self.connection.cursor()
-
-        self.connection.commit()
 
     def url_to_html(self, url, params, headers = None):
         """Extract Web content into an HTML tree object
